MERO-422: rc == 0 at ut_backend_init_cfg() be/ut/helper.c:353
m0_stob_domain_init() invoked by be_domain_start_normal_pre() fails with -ENOENT.

* reqh-ut, fol-ut, balloc-ut fail if executed out of order with other uts.
  e.g. sudo scripts/m0 run-ut -t dtm-transmit-ut,fol-ut

* m0_be_ut_backend_init_cfg() uses local static boolean variable mkfs_executed to
  decide if mkfs needs to be performed.
  Being a static variable (mkfs_executed), if set "true" by one ut, stays true for next ut.
  m0_be_ut_backend_init() invokes m0_be_ut_backend_init_cfg() with default mkfs == false.
  This problem arises if one ut invokes m0_be_ut_backend_init() with non-default stob location
  (struct m0_be_ut_backend::but_stob_domain_location != NULL) and other ut uses default
  stob domain location (struct m0_be_ut_backend::but_stob_domain_location == NULL).
  "mkfs_executed" is set regardless of struct m0_be_ut_backend::but_stob_domain_location value.
  Thus on invocation of m0_be_ut_backend_init() by subsequent ut, causes above failure as
  mkfs is not performed.

+++ be/ut/helper.c
@@ -370,7 +370,12 @@ M0_INTERNAL void m0_be_ut_backend_init_cfg(struct m0_be_ut_backend *ut_be,
 {
 	static bool mkfs_executed = false;
 
-	if (!mkfs_executed && cfg == NULL) {
+	/*
+	 * Set mkfs_executed iff default stob domain location is used,
+	 * i.e. ut_be->but_stob_domain_location == NULL.
+	 */
+	if (!mkfs_executed && cfg == NULL &&
+	    ut_be->but_stob_domain_location == NULL) {
 		mkfs = true;
 		mkfs_executed = true;
 	}
